requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.1
Django==5.1
django-allauth==64.1.0
django-crispy-forms==2.3
django-widget-tweaks==1.5.0
pillow==10.4.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
